requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pillow
scikit-learn
numpy
nltk